+++ docs/changelog.rst
@@ -8,6 +8,7 @@ CHANGELOG
 **New features**
 
 - APIv2 : Add 'children' and 'parent' fields to Outdoor Site serialization
+- APIv2 : Add filter by pratices on outdoor courses
 
 
 2.66.0 (2021-09-27)